app/requirements.txt
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
confluent-kafka==2.4.0
Faker==30.3.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.1
orjson==3.10.7
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
quixstreams==3.0.0
referencing==0.35.1
requests==2.32.3
rocksdict==0.3.23
rpds-py==0.20.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
